requirements.txt
streamlit
numpy
pillow
cv2
google-generativeai
pyttsx3